requirements.txt
pytest==8.3.4
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-xdist==3.6.1
allure-pytest==2.13.5
playwright==1.49.1